+++ Configuration/Icons.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 use TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider;
 
 return [
-    'viewmend-module' => [
+    'viewmend-product-mark' => [
         'provider' => SvgIconProvider::class,
-        'source' => 'EXT:viewmend_core/Resources/Public/Icons/module-viewmend.svg',
+        'source' => 'EXT:viewmend_core/Resources/Public/Icons/viewmend-product-mark.svg',
     ],
 ];